breach/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.36
pydantic==2.11.7
python-dotenv==1.0.0
redis==5.3.1
bcrypt==4.2.0
requests==2.31.0
python-jose==3.3.0
apscheduler==3.10.1
aiosmtplib==2.0.1
